:construction: remove old mock data, move mock data generation to knex seeds folder, update mock data generation to match table schema
