--- src/mock/small/users.json
@@ -1,142 +1,0 @@
-[
-  {
-    "id": "fd7398ab-7968-4a1c-adfb-142d30b6a936",
-    "firstName": "Caroline",
-    "lastName": "Swift",
-    "phoneNumber": "1-446-490-1781 x3745",
-    "email": "Caroline_Swift50@hotmail.com"
-  },
-  {
-    "id": "24e1899b-101d-4580-acd8-70fc97c889a2",
-    "firstName": "Meredith",
-    "lastName": "O'Reilly",
-    "phoneNumber": "(467) 651-1451 x721",
-    "email": "Meredith_OReilly13@hotmail.com"
-  },
-  {
-    "id": "a109278c-9f00-4d88-a5e6-0ac80925eacb",
-    "firstName": "Raul",
-    "lastName": "Dare",
-    "phoneNumber": "(479) 445-7179 x803",
-    "email": "Raul.Dare@hotmail.com"
-  },
-  {
-    "id": "23c9802c-90c7-4035-9477-c75efa93e418",
-    "firstName": "Ansel",
-    "lastName": "Buckridge",
-    "phoneNumber": "859.884.1524 x865",
-    "email": "Ansel.Buckridge@gmail.com"
-  },
-  {
-    "id": "2b6221e1-4141-4c83-98b2-8efbb0a7e355",
-    "firstName": "Laury",
-    "lastName": "Torphy",
-    "phoneNumber": "566.325.2621 x843",
-    "email": "Laury_Torphy@yahoo.com"
-  },
-  {
-    "id": "def2c9e2-63b6-4fe8-adfb-6cff2643bbd3",
-    "firstName": "Brandy",
-    "lastName": "Runolfsson",
-    "phoneNumber": "415.916.5708 x4878",
-    "email": "Brandy.Runolfsson@gmail.com"
-  },
-  {
-    "id": "9ba59fbf-e435-4ef8-9a6a-f5f56f87d9cc",
-    "firstName": "Rowena",
-    "lastName": "Kunze",
-    "phoneNumber": "(291) 584-9813 x308",
-    "email": "Rowena_Kunze20@yahoo.com"
-  },
-  {
-    "id": "723fc0c9-78df-479c-bf08-985ee20446ad",
-    "firstName": "Dolores",
-    "lastName": "Boyle",
-    "phoneNumber": "278-889-5201 x55622",
-    "email": "Dolores.Boyle53@gmail.com"
-  },
-  {
-    "id": "c9d4b61a-679b-4a8c-b913-2ec7bdc1c3ff",
-    "firstName": "Brisa",
-    "lastName": "Bartoletti",
-    "phoneNumber": "(358) 434-5843 x2649",
-    "email": "Brisa_Bartoletti@gmail.com"
-  },
-  {
-    "id": "52a7b64e-804a-42c2-95ad-3a8e0b71630c",
-    "firstName": "Dagmar",
-    "lastName": "Hirthe",
-    "phoneNumber": "295.804.8001 x811",
-    "email": "Dagmar_Hirthe44@yahoo.com"
-  },
-  {
-    "id": "1157e469-98cb-4bc6-b16c-3b614a12272f",
-    "firstName": "Lydia",
-    "lastName": "Franey",
-    "phoneNumber": "397-484-2568 x156",
-    "email": "Lydia.Franey@gmail.com"
-  },
-  {
-    "id": "c9d12b2f-aca2-4eee-9538-219aa97a3746",
-    "firstName": "Margarett",
-    "lastName": "Torp",
-    "phoneNumber": "(329) 379-1021 x09442",
-    "email": "Margarett_Torp@yahoo.com"
-  },
-  {
-    "id": "067ac486-eb4c-441f-95cf-6c6634eab698",
-    "firstName": "Alexa",
-    "lastName": "Howell",
-    "phoneNumber": "321-444-6604 x499",
-    "email": "Alexa_Howell98@hotmail.com"
-  },
-  {
-    "id": "8379e394-8d2f-4616-895f-3ae956cdc6e7",
-    "firstName": "Alessandro",
-    "lastName": "Williamson",
-    "phoneNumber": "555.717.7803 x853",
-    "email": "Alessandro.Williamson92@yahoo.com"
-  },
-  {
-    "id": "cfc0ba28-e4b1-4416-9512-144d2ab524df",
-    "firstName": "Pascale",
-    "lastName": "Murphy",
-    "phoneNumber": "1-830-796-0750",
-    "email": "Pascale_Murphy19@hotmail.com"
-  },
-  {
-    "id": "ed63cd7d-bdd3-4ec5-becf-14d2470dd805",
-    "firstName": "Jude",
-    "lastName": "Kuhic",
-    "phoneNumber": "(443) 795-0158 x7718",
-    "email": "Jude_Kuhic6@gmail.com"
-  },
-  {
-    "id": "c12f95dd-9446-48bc-b124-9e0142d53820",
-    "firstName": "Owen",
-    "lastName": "Towne",
-    "phoneNumber": "561-250-5099 x94990",
-    "email": "Owen_Towne@yahoo.com"
-  },
-  {
-    "id": "95e624f2-fdfb-45ca-bbd5-ee0fce02eb6d",
-    "firstName": "Verla",
-    "lastName": "Sawayn",
-    "phoneNumber": "(510) 837-6790 x3618",
-    "email": "Verla_Sawayn66@hotmail.com"
-  },
-  {
-    "id": "1138f46e-b5d8-4c25-a1d4-fc01e1a5d3b0",
-    "firstName": "Cameron",
-    "lastName": "Shields",
-    "phoneNumber": "775.401.4006 x71935",
-    "email": "Cameron.Shields71@hotmail.com"
-  },
-  {
-    "id": "6dee7d92-9e43-4e20-9a41-8b815fb812a3",
-    "firstName": "Jared",
-    "lastName": "Harber",
-    "phoneNumber": "998.845.6845 x38703",
-    "email": "Jared_Harber@hotmail.com"
-  }
-]
